Split codebase into two folders (see `/docs/ROADMAP.md`)
