+++ interpreter/core/code_interpreters/languages/applescript.py
