+++ interpreter/core/code_interpreters/languages/python_vision.py
