+++ interpreter/core/code_interpreters/languages/shell.py
