+++ interpreter/core/utils/get_user_info_string.py
