+++ interpreter/core/utils/merge_deltas.py
