+++ interpreter/core/utils/truncate_output.py
