+++ interpreter/terminal_interface/conversation_navigator.py
@@ -9,9 +9,9 @@ import subprocess
 
 import inquirer
 
-from ..utils.display_markdown_message import display_markdown_message
-from ..utils.local_storage_path import get_storage_path
 from .render_past_conversation import render_past_conversation
+from .utils.display_markdown_message import display_markdown_message
+from .utils.local_storage_path import get_storage_path
 
 
 def conversation_navigator(interpreter):
