+++ interpreter/terminal_interface/render_past_conversation.py
@@ -1,6 +1,6 @@
-from ..utils.display_markdown_message import display_markdown_message
 from .components.code_block import CodeBlock
 from .components.message_block import MessageBlock
+from .utils.display_markdown_message import display_markdown_message
 
 
 def render_past_conversation(messages):
