+++ interpreter/terminal_interface/utils/find_image_path.py
