+++ pyproject.toml
@@ -42,7 +42,7 @@ requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-interpreter = "interpreter:cli"
+interpreter = "interpreter:start_terminal_interface"
 
 [tool.poetry.extras]
 safe = ["semgrep"]
